requirements.txt
astroid==2.4.2
grpcio==1.30.0
grpcio-tools==1.30.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
protobuf==3.12.2
pylint==2.5.3
six==1.15.0
toml==0.10.1
wrapt==1.12.1
